Updated SignUpForm to use external validators
